Split out deadzones into another module
