+++ src/plugin.rs
@@ -181,12 +181,12 @@ impl<A: Actionlike + TypePath> Plugin for InputManagerPlugin<A> {
             .register_type::<MouseWheelDirection>()
             .register_type::<MouseMotionDirection>()
             // Input processors and settings
-            .register_type::<InputClamp>()
-            .register_type::<InputNormalizer>()
+            .register_type::<ValueLimit>()
+            .register_type::<ValueNormalizer>()
             .register_type::<SingleAxisSettings>()
-            .register_type::<SingleAxisDeadzone>()
+            .register_type::<Deadzone1>()
             .register_type::<DualAxisSettings>()
-            .register_type::<DualAxisDeadzone>()
+            .register_type::<Deadzone2>()
             // Resources
             .init_resource::<ToggleActions<A>>()
             .init_resource::<ClashStrategy>();
